requirements.txt
openai==0.27.8
python-dotenv==1.0.0
yaspin==2.3.0
bs4==0.0.1
user-agent==0.1.10
fake-useragent==1.1.3
tiktoken==0.4.0
sqlmodel
flask_socketio==5.3.2
flask_session
flask==2.2.2
flask_cors
redis
simple-websocket==0.10.1
python-engineio==4.4.0
python-socketio==5.8.0
langdetect
python-iso639
